Move views to its own directory
