+++ web/img-mngr/src/components/views/list-items.html
