+++ web/img-mngr/src/components/views/path-media.html
