+++ web/img-mngr/src/components/views/root-media.html
